requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
discord==1.0.1
discord.py==1.7.2
idna==3.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
pylint==2.6.0
python-dotenv==0.17.1
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
requests